requirements.txt
Flask==2.2.2
transformers==4.26.1
torch==1.13.1
faiss-cpu==1.7.3
